@@ -17,8 +17,9 @@
   "author": "Lakshita Ganesh Kumar",
   "license": "MIT",
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^7.6.3",
     "react-scripts": "^5.0.1"
   },
   "homepage": "http://lakshitaganeshkumar.github.io/20th-year-vision-board",
